refactor(records): move five surfaces out of RecordPatches.cs into partials
The if-chain batch-sibling-issues-by-file.md names as a standing conflict point now
sits alone in RecordPatches.DataAccessDispatch.cs, away from Register(). Pure move:
every non-blank line of the original is present exactly once across the family.
41 build warnings before and after.
