+++ AlRunner/Patches/RecordPatches.cs
@@ -935,58 +935,6 @@ public static partial class RecordPatches
     }
 
     /// <summary>
-    /// Replacement for NavSession.Database getter.
-    /// NavSession.Database => Tenant.Database which requires a real tenant.
-    /// Return the skeleton NavDatabase instead.
-    /// </summary>
-    [MethodImpl(MethodImplOptions.NoInlining)]
-    public static object? NavSession_get_Database(object self)
-    {
-        return _skeletonDatabase;
-    }
-
-    // Cached NavTenant.database (LazyEx<NavDatabase>) field, resolved lazily.
-    private static FieldInfo? _fNavTenantDatabase;
-    private static bool _fNavTenantDatabaseResolved;
-
-    /// <summary>
-    /// Replacement for NavTenant.get_Database. The real getter throws
-    /// ArgumentNullException("NavDatabase") when the tenant's `database` LazyEx is null,
-    /// which it always is on the skeleton (MetadataPatches leaves it null by design).
-    /// Under R2R, `NavSession.Database => Tenant.Database` is inlined past our
-    /// NavSession.get_Database redirect, so callers like ALNavApp.ALGetModuleInfo
-    /// (FeatureTelemetry.LogUsage during Purch.-Post) reach NavTenant.Database directly.
-    /// Return the runner's skeleton NavDatabase (which carries collation/sorting and an
-    /// empty-family SqlDatabaseProperties) instead of throwing. If a real database LazyEx
-    /// is ever present we honour it. Runtime-engine layer; faithful — the skeleton DOES
-    /// have a minimal database, returning it is more correct than throwing.
-    /// </summary>
-    [MethodImpl(MethodImplOptions.NoInlining)]
-    public static object? NavTenant_get_Database(object self)
-    {
-        if (!_fNavTenantDatabaseResolved)
-        {
-            _fNavTenantDatabase = self?.GetType().GetField("database",
-                BindingFlags.NonPublic | BindingFlags.Instance)
-                ?? self?.GetType().BaseType?.GetField("database",
-                    BindingFlags.NonPublic | BindingFlags.Instance);
-            _fNavTenantDatabaseResolved = true;
-        }
-        if (self != null && _fNavTenantDatabase != null)
-        {
-            var lazy = _fNavTenantDatabase.GetValue(self);
-            if (lazy != null)
-            {
-                // LazyEx<NavDatabase>.Value — return the real DB if the tenant has one.
-                var pValue = lazy.GetType().GetProperty("Value");
-                var v = pValue?.GetValue(lazy);
-                if (v != null) return v;
-            }
-        }
-        return _skeletonDatabase;
-    }
-
-    /// <summary>
     /// Replacement for TempTableDataProvider.ctor(NavSession, NCLMetaTable).
     /// The real ctor calls navSession.Database.CollationAwareStringComparer which NREs on our
     /// skeleton session (no Tenant). We manually set all fields instead.
@@ -1000,614 +948,6 @@ public static partial class RecordPatches
         var pkSortingFields = _pNclMetaKeySortingFieldsWithPK?.GetValue(table.PrimaryKey);
         _fTtdpPrimaryKeySortingFields?.SetValue(self, pkSortingFields);
         _fTtdpComparer?.SetValue(self, _collationComparer);
-    }
-
-    /// <summary>
-    /// Replacement for TempTableDataProvider.CalcNumeric(CalcNumericProviderRequest).
-    /// The real override throws NotSupportedException; this replacement iterates in-memory rows
-    /// via the private Filter() helper and aggregates each requested FlowField.
-    /// <para>#2937 — the doc comment here used to claim count/sum/average were "the only three
-    /// calculation methods routed through CalcNumeric", and the result switch ended in
-    /// <c>_ =&gt; sums[j]</c>, so Min/Max/Lookup/Exists/None were all answered with the SUM
-    /// accumulator. For Min/Max nothing ever wrote that accumulator, so they came back as a
-    /// constant 0 whatever the data — right for an empty source set only by coincidence.
-    /// Count/Sum/Average is indeed what BC ROUTES here (DistinctSourceTable.AddField buckets
-    /// Min/Max into MinMaxFlowFields → CalcMinMax, Lookup and Exists into their own lists), but
-    /// "BC does not send it" is not a reason to answer it wrongly: Min/Max are now aggregated
-    /// properly and everything else throws. Aggregation itself lives in
-    /// <see cref="ComputeCalcNumericAggregate"/>.</para>
-    /// <para>NegateResult (<c>CalcFormula = -sum(...)</c>) is applied here because BC applies it
-    /// at this level too: NavSqlAggregateCommand's aggregate reader negates each aggregated
-    /// value inside the provider, before the FieldDictionary is returned. The negation itself is
-    /// BC's own NCLMetaCalculationFormula.NegateValue, reached through
-    /// <see cref="FlowFieldPatches.NegateAggregateResult"/> so the two runner paths that negate
-    /// a FlowField aggregate share one implementation (#1708, #2323).</para>
-    /// <para>REACHABILITY, measured rather than assumed (#2937 left this open): instrumenting
-    /// this method's result loop and running the whole al-language corpus (2496 tests) plus all
-    /// of tests/runner-extras (264 tests) produced ZERO hits. FlowFieldPatches hooks
-    /// NavRecord.CalcFieldsAsync ahead of FlowFieldsHelper, so ordinary CalcFields — on a
-    /// temporary record too, since every runner table is backed by TempTableDataProvider — never
-    /// arrives here. That is why the coverage for this method is C# (CalcNumericAggregateTests)
-    /// and not AL: there is no AL shape today that reaches it, so an AL test would pass without
-    /// executing a line of it.</para>
-    /// </summary>
-    [MethodImpl(MethodImplOptions.NoInlining)]
-    public static object TempTableDataProvider_CalcNumeric(object self, object request)
-    {
-        // No Date (2000000007) materialisation here since #3506: BC's own DateDataProvider
-        // answers that table, so a Date read never reaches a TempTableDataProvider.
-
-        var rt = request.GetType();
-        var companyToken   = (int)BcShape.Property(rt, "CompanyToken", "AL record data access").GetValue(request)!;
-        var filtersAndMarks = BcShape.Property(rt, "FiltersAndMarks", "AL record data access").GetValue(request);
-        var sourceTable    = (NCLMetaTable)BcShape.Property(
-            rt, "MetaApplicationObject", "AL record data access").GetValue(request)!;
-        var fieldsToCalc   = (FieldList)BcShape.Property(
-            rt, "FieldsToCalculate", "AL record data access").GetValue(request)!;
-        int fieldCount     = fieldsToCalc.Count;
-
-        var primaryKeySortingFields = _fTtdpPrimaryKeySortingFields!.GetValue(self);
-        int recordCount = 0;
-
-        // Per-field state resolved once, not once per row: the calculation method, the source
-        // field the formula names (Count has none), and the source values collected across the
-        // matching rows. Only the value-consuming methods get a list — a Count field's slot
-        // stays null and nothing is collected for it.
-        //
-        // Collecting the values rather than folding them into a running total is the same shape
-        // RecordPatches.QueryProjection's ComputeAggregateCore uses, and it is what lets one
-        // helper own Sum/Average/Min/Max instead of the result switch reading whichever
-        // accumulator happened to be written. The cost is one NavValue REFERENCE per matching
-        // row per aggregated field — the rows' own value objects, not copies, and the rows are
-        // already materialised in the temp store by the Filter() call below.
-        var methods = new NCLMetaCalculationMethod[fieldCount];
-        var sourceFields = new NCLMetaField?[fieldCount];
-        var sourceValues = new List<NavValue?>?[fieldCount];
-        for (int i = 0; i < fieldCount; i++)
-        {
-            var field = (NCLMetaField)fieldsToCalc[i];
-            methods[i] = field.CalculationFormula.CalculationMethod;
-            if (methods[i] is NCLMetaCalculationMethod.Sum or NCLMetaCalculationMethod.Average
-                or NCLMetaCalculationMethod.Min or NCLMetaCalculationMethod.Max)
-            {
-                sourceFields[i] = sourceTable.GetFieldByNo(field.CalculationFormula.FieldId, trapError: true);
-                // Source field unresolvable → no values are collected and the aggregate answers
-                // its empty-set value, which is what this method did before #2937 and what
-                // FlowFieldPatches' own srcFieldColumn < 0 arm does. Left as-is deliberately:
-                // making it loud is a change to the shared FlowField behaviour, not to this
-                // method, and belongs with the sibling rather than half-applied here.
-                if (sourceFields[i] != null) sourceValues[i] = new List<NavValue?>();
-            }
-        }
-
-        var rows = (System.Collections.IEnumerable)_mTtdpFilter!.Invoke(
-            self, new object?[] { companyToken, filtersAndMarks, null, primaryKeySortingFields, false })!;
-
-        foreach (TempTableRecordBuffer row in rows)
-        {
-            checked { recordCount++; }
-            for (int i = 0; i < fieldCount; i++)
-            {
-                var values = sourceValues[i];
-                if (values == null) continue;
-                values.Add(row[sourceFields[i]!.ColumnIndex]);
-            }
-        }
-
-        var tuples = new Tuple<INavFieldMetadata, NavValue>[fieldCount];
-        for (int j = 0; j < fieldCount; j++)
-        {
-            var field = (NCLMetaField)fieldsToCalc[j];
-            var formula = field.CalculationFormula;
-
-            // BC negates at this level too — see the method's doc comment. The negation is
-            // passed IN rather than applied after the call so that "aggregate, then negate iff
-            // NegateResult" is one decision with one owner, and so a test can drive it: BC's
-            // own NCLMetaCalculationFormula.NegateValue resolves SourceField through the
-            // metadata registry and therefore needs a live session, which a unit test does not
-            // have. Exists never reaches the negation (ComputeCalcNumericAggregate throws for
-            // it first), so the #2323 exist carve-out in FlowFieldPatches has no counterpart to
-            // make here.
-            var navValue = ComputeCalcNumericAggregate(
-                methods[j], field, recordCount,
-                (IEnumerable<NavValue?>?)sourceValues[j] ?? Array.Empty<NavValue?>(),
-                $"{field.Parent?.TableName}.{field.FieldName}",
-                formula.NegateResult,
-                v => FlowFieldPatches.NegateAggregateResult(
-                    formula, v, "TempTableDataProvider.CalcNumeric"));
-
-            tuples[j] = new Tuple<INavFieldMetadata, NavValue>(field, navValue);
-        }
-
-        return _ctorFieldDictionaryNavValue!.Invoke(new object[] { tuples })!;
-    }
-
-    /// <summary>
-    /// One CalcNumeric field's aggregate, over the source values of the rows that matched the
-    /// formula's filters (<paramref name="rowCount"/> is how many rows matched — Count's answer,
-    /// and Average's divisor, both of which count rows rather than non-null values).
-    /// <para>Min/Max reuse <see cref="FlowFieldPatches.NavValueCompare"/> and
-    /// <see cref="FlowFieldPatches.TypedDefaultForField"/> rather than re-deriving comparison or
-    /// default semantics — the same reuse RecordPatches.QueryProjection's ComputeAggregateCore
-    /// makes, so all three of the runner's aggregate paths order values and answer an empty set
-    /// identically.</para>
-    /// <para>The empty-source-set answers are deliberate, not a fallthrough: corpus PR
-    /// StefanMaron/BusinessCentral.AL.Language.Tests#171 measured real BC on eight service tiers
-    /// answering 0 for min/max/average over no matching rows — and 0D for a Date-typed one,
-    /// which is why the answer is the field's OWN typed default and not a numeric zero.</para>
-    /// <para>Anything else throws. BC never routes Exists/Lookup/None through CalcNumeric
-    /// (DistinctSourceTable.AddField buckets them into their own field lists), so one arriving
-    /// here means the dispatch changed — and per loud-failures.md that has to be loud rather
-    /// than answered with a default. Min and Max are answered even though BC does not route
-    /// them here either, because the aggregation is exactly the same work and a wrong constant
-    /// was the #2937 defect.</para>
-    /// </summary>
-    /// <param name="negateResult">the formula's <c>NegateResult</c> — the leading minus in
-    /// <c>CalcFormula = -sum(...)</c> (#1708).</param>
-    /// <param name="negate">applies that minus, and is only ever called when
-    /// <paramref name="negateResult"/> is true. Required rather than optional: a null default
-    /// would let a caller silently answer the POSITIVE aggregate for a negated formula, which
-    /// is the exact silent wrong value #1708 is about. Production passes
-    /// <see cref="FlowFieldPatches.NegateAggregateResult"/>, so BC's own
-    /// NCLMetaCalculationFormula.NegateValue stays the single owner of the negation.</param>
-    internal static NavValue ComputeCalcNumericAggregate(
-        NCLMetaCalculationMethod method,
-        INavValueMetadata resultMetadata,
-        int rowCount,
-        IEnumerable<NavValue?> sourceValues,
-        string fieldDescription,
-        bool negateResult,
-        Func<NavValue, NavValue> negate)
-    {
-        if (negateResult && negate == null)
-            throw new ArgumentNullException(nameof(negate),
-                $"NegateResult is set on '{fieldDescription}' but no negation was supplied — "
-                + "answering the positive aggregate would be the silent wrong value of #1708");
-
-        var aggregate = ComputeCalcNumericAggregateCore(
-            method, resultMetadata, rowCount, sourceValues, fieldDescription);
-        return negateResult ? negate(aggregate) : aggregate;
-    }
-
-    private static NavValue ComputeCalcNumericAggregateCore(
-        NCLMetaCalculationMethod method,
-        INavValueMetadata resultMetadata,
-        int rowCount,
-        IEnumerable<NavValue?> sourceValues,
-        string fieldDescription)
-    {
-        switch (method)
-        {
-            case NCLMetaCalculationMethod.Count:
-                return NavValue.CreateNavValueFromObject(resultMetadata, rowCount);
-
-            case NCLMetaCalculationMethod.Sum:
-            case NCLMetaCalculationMethod.Average:
-            {
-                Decimal18 sum = default;
-                foreach (var v in sourceValues)
-                {
-                    if (v == null) continue;
-                    sum = checked(sum + v.ToDecimal());
-                }
-                if (method == NCLMetaCalculationMethod.Sum)
-                    // An empty sum is 0 by arithmetic, not by accumulator accident.
-                    return NavValue.CreateNavValueFromObject(resultMetadata, sum);
-                return rowCount > 0
-                    ? NavValue.CreateNavValueFromObject(resultMetadata, sum / rowCount)
-                    : EmptyAggregateDefault(resultMetadata);
-            }
-
-            case NCLMetaCalculationMethod.Min:
-            case NCLMetaCalculationMethod.Max:
-            {
-                NavValue? best = null;
-                foreach (var v in sourceValues)
-                {
-                    if (v == null) continue;
-                    if (best == null
-                        || (method == NCLMetaCalculationMethod.Min && FlowFieldPatches.NavValueCompare(v, best) < 0)
-                        || (method == NCLMetaCalculationMethod.Max && FlowFieldPatches.NavValueCompare(v, best) > 0))
-                        best = v;
-                }
-                return best ?? EmptyAggregateDefault(resultMetadata);
-            }
-
-            default:
-                throw new RunnerOutOfScopeException(
-                    "TempTableDataProvider.CalcNumeric",
-                    $"not-yet-implemented — CalculationMethod {method} on '{fieldDescription}' "
-                    + "is not aggregated by CalcNumeric. BC routes Exists to CalcExists and "
-                    + "Lookup to CalcLookup (DistinctSourceTable.AddField), so a CalcNumeric "
-                    + "request carrying one means the dispatch changed; answering it with the "
-                    + "sum accumulator was issue #2937",
-                    "todo");
-        }
-    }
-
-    /// <summary>
-    /// What an aggregate answers when no row contributed a value: the result field's OWN typed
-    /// default (0 for Decimal/Integer, 0D for Date, …), never a bare numeric literal — same
-    /// chain FlowFieldPatches' Min/Max/Lookup arms use.
-    /// </summary>
-    private static NavValue EmptyAggregateDefault(INavValueMetadata resultMetadata)
-        => FlowFieldPatches.TypedDefaultForField(resultMetadata)
-           ?? NavValue.CreateNavValueFromObject(resultMetadata, 0);
-
-    /// Pre-populate the skeleton session's dataAccessSource field directly.
-    /// NavSession.DataAccessSource getter is a trivial field return and gets inlined by JIT,
-    /// so the JMP hook on it never fires. We must inject the DAS into the field directly.
-    /// </summary>
-    public static void InitializeSkeletonSession(object skeletonSession)
-    {
-        Console.Error.WriteLine($"[RecordPatches] InitializeSkeletonSession: _fSessionDataAccessSource={_fSessionDataAccessSource != null}, _tDataAccessSource={_tDataAccessSource != null}");
-        if (_fSessionDataAccessSource == null || _tDataAccessSource == null) return;
-
-        // If already set, nothing to do.
-        var existing = _fSessionDataAccessSource.GetValue(skeletonSession);
-        Console.Error.WriteLine($"[RecordPatches] existing DAS on session: {existing}");
-        if (existing != null) return;
-
-        // Ensure skeleton DB (needed by TempTableDataProvider ctor via navSession.Database.CollationAwareStringComparer)
-        EnsureSkeletonDatabase(skeletonSession);
-
-        // Build skeleton DataAccessSource
-        var das = RuntimeHelpers.GetUninitializedObject(_tDataAccessSource);
-        _fDasSession!.SetValue(das, skeletonSession);
-        _fDasGlobalFilters!.SetValue(das, Activator.CreateInstance(_tGlobalFilters!));
-        _fDasTableVersionTokens!.SetValue(das, _mCreateForTempTable!.Invoke(null, null));
-
-        // Pre-populate sessionTransactionManager so the lazy-init path in
-        // DataAccessSource.get_SessionTransactionManager (which would otherwise call
-        // CreateAppDataAccess → CreateAppDataProvider → NRE) is bypassed. The skeleton STM
-        // carries a single LogicalTransaction with TransactionType=Update so that the real
-        // BC body of ALDatabase.get_ALCurrentTransactionType returns Update without machinery,
-        // and SessionTransactionManager.AnyHasWriteTransactionStarted returns false (empty
-        // transactionManagers dict — no per-table TM has ever begun a write transaction).
-        // Faithful: the runner has no real write transaction system; Update is BC's default
-        // for "browsing/reading without a lock-mode override", and IsInWriteTransaction is
-        // observably false because nothing has called BeginTransaction.
-        if (_skeletonSessionTransactionManager != null && _fDasSessionTransactionManager != null)
-            _fDasSessionTransactionManager.SetValue(das, _skeletonSessionTransactionManager);
-
-        // Inject directly into the session field (bypass the inlined getter)
-        _fSessionDataAccessSource.SetValue(skeletonSession, das);
-        Console.Error.WriteLine($"[RecordPatches] Skeleton DAS injected on session: {das.GetType().Name}");
-
-        // Build a minimal NavSystemCodeunitFactory+GlobalTriggers on the skeleton company so that
-        // NavRecord.IsGlobalTriggerImplemented doesn't NRE when it calls
-        // Session.SystemCodeunitFactory.GlobalTriggers.GetTriggersOnTable().
-        // The factory's GlobalTriggers.session is our skeleton which is not "IsCompanyOpen",
-        // so GetTriggersOnTable() returns Triggers.None immediately.
-        InjectSkeletonSystemCodeunitFactory(skeletonSession);
-
-        // Populate the auto-property backing field for NavSession.ErrorCollection.
-        // The real auto-property `internal ErrorCollection ErrorCollection { get; } = new ErrorCollection();`
-        // is initialised in NavSession's instance ctor — which doesn't run for our
-        // RuntimeHelpers.GetUninitializedObject skeleton. Without this, NavMethodScope.RunBehaviorAsync
-        // NREs the moment any AL method tagged [ErrorBehavior(Collect)] runs (via
-        // session.ErrorCollection.StartCollecting()). Construct the real ErrorCollection so
-        // StartCollecting / StopCollecting / Collect all execute unmodified BC code (Option C —
-        // reuse service-tier code per HANDOFF §2 invariant 4).
-        InjectSkeletonErrorCollection(skeletonSession);
-    }
-
-    /// <summary>
-    /// Build a skeleton SessionTransactionManager whose two read-only getters resolve to
-    /// BC-faithful defaults for a session with no real write transaction:
-    ///   * CurrentTransactionType => TransactionType.Update (BC's default lock mode)
-    ///   * AnyHasWriteTransactionStarted => false (empty per-table TM dict)
-    /// Built reflectively via RuntimeHelpers.GetUninitializedObject so we can skip the real
-    /// ctor (which wires TransactionManagers tied to TenantDataAccess / AppDataAccess — both
-    /// of which require a real connection in our skeleton). All fields the two getters read
-    /// are populated explicitly; nothing else is touched.
-    /// </summary>
-    private static object? BuildSkeletonSessionTransactionManager()
-    {
-        var nclAsm = AppDomain.CurrentDomain.GetAssemblies()
-            .FirstOrDefault(a => a.GetName().Name == "Microsoft.Dynamics.Nav.Ncl");
-        var typesAsm = AppDomain.CurrentDomain.GetAssemblies()
-            .FirstOrDefault(a => a.GetName().Name == "Microsoft.Dynamics.Nav.Types");
-        if (nclAsm == null || typesAsm == null) return null;
-
-        var tStm = nclAsm.GetType("Microsoft.Dynamics.Nav.Runtime.SessionTransactionManager");
-        var tTm = nclAsm.GetType("Microsoft.Dynamics.Nav.Runtime.TransactionManager");
-        var tLt = tTm?.GetNestedType("LogicalTransaction", BindingFlags.NonPublic | BindingFlags.Public);
-        var tTrType = typesAsm.GetType("Microsoft.Dynamics.Nav.Types.TransactionType");
-        if (tStm == null || tTm == null || tLt == null || tTrType == null) return null;
-
-        // LogicalTransaction is a parameterless internal type whose backing fields are
-        // exactly the auto-properties. Construct it and set TransactionType = Update (ordinal 1).
-        var lt = Activator.CreateInstance(tLt, nonPublic: true);
-        if (lt == null) return null;
-        var fLtType = tLt.GetField("<TransactionType>k__BackingField",
-            BindingFlags.NonPublic | BindingFlags.Instance);
-        if (fLtType == null) return null;
-        var updateValue = Enum.ToObject(tTrType, 1); // TransactionType.Update
-        fLtType.SetValue(lt, updateValue);
-
-        // Build a TransactionManager via GetUninitializedObject and populate just the
-        // logicalTransactions stack so get_CurrentTransactionType (which Peek()s the stack)
-        // returns Update without going through TransactionalDataProvider / NavSession state.
-        var tm = RuntimeHelpers.GetUninitializedObject(tTm);
-        var fTmLogicalTransactions = tTm.GetField("logicalTransactions",
-            BindingFlags.NonPublic | BindingFlags.Instance);
-        if (fTmLogicalTransactions == null) return null;
-        var stackType = typeof(Stack<>).MakeGenericType(tLt);
-        var stack = Activator.CreateInstance(stackType)!;
-        stackType.GetMethod("Push")!.Invoke(stack, new[] { lt });
-        fTmLogicalTransactions.SetValue(tm, stack);
-
-        // Build the SessionTransactionManager and populate:
-        //   defaultTransactionManager → our skeleton TM (so STM.CurrentTransactionType works)
-        //   transactionManagers       → empty dict (so AnyHasWriteTransactionStarted returns false)
-        var stm = RuntimeHelpers.GetUninitializedObject(tStm);
-        var fStmDefault = tStm.GetField("defaultTransactionManager",
-            BindingFlags.NonPublic | BindingFlags.Instance);
-        var fStmDict = tStm.GetField("transactionManagers",
-            BindingFlags.NonPublic | BindingFlags.Instance);
-        fStmDefault?.SetValue(stm, tm);
-        if (fStmDict != null)
-        {
-            // dict type: ConcurrentDictionary<TransactionManagerKey, TransactionManager>
-            var dict = Activator.CreateInstance(fStmDict.FieldType);
-            fStmDict.SetValue(stm, dict);
-        }
-
-        return stm;
-    }
-
-    private static void InjectSkeletonErrorCollection(object skeletonSession)
-    {
-        var nclAsm = AppDomain.CurrentDomain.GetAssemblies()
-            .FirstOrDefault(a => a.GetName().Name == "Microsoft.Dynamics.Nav.Ncl");
-        if (nclAsm == null) return;
-        var tErrorCollection = nclAsm.GetType("Microsoft.Dynamics.Nav.Runtime.ErrorCollection");
-        if (tErrorCollection == null) return;
-
-        var fErrorCollection = skeletonSession.GetType().GetField("<ErrorCollection>k__BackingField",
-            BindingFlags.NonPublic | BindingFlags.Instance);
-        if (fErrorCollection == null)
-        {
-            Console.Error.WriteLine("[RecordPatches] ErrorCollection backing field not found");
-            return;
-        }
-        // Already populated? leave alone.
-        if (fErrorCollection.GetValue(skeletonSession) != null)
-        {
-            WireNavCurrentThreadSession(skeletonSession);
-            return;
-        }
-
-        // ErrorCollection has only field initialisers (collectedErrors=null, currentCollectionScopeStart=-1)
-        // and a default ctor; Activator.CreateInstance runs the field-initialiser block.
-        var ec = Activator.CreateInstance(tErrorCollection, nonPublic: true);
-        fErrorCollection.SetValue(skeletonSession, ec);
-        Console.Error.WriteLine("[RecordPatches] Skeleton ErrorCollection injected");
-
-        WireNavCurrentThreadSession(skeletonSession);
-    }
-
-    /// <summary>
-    /// Wire NavCurrentThread.Session to return _skeletonSession. NavCurrentThread.Session reads
-    /// NavThreadLocalStorage.Current.Session?.Target — an AsyncLocal&lt;IReference&lt;NavSession&gt;&gt;.
-    /// Setting it on the bootstrap thread propagates via ExecutionContext into the test threads.
-    /// Without this, ALIsCollectingErrors / ALHasCollectedErrors / ALClearCollectedErrors /
-    /// ALGetCollectedErrors all dereference NavCurrentThread.Session (null) → NRE; the AL tests
-    /// that rely on the [ErrorBehavior(Collect)] surface (CollectThenClear, ClearCollectedErrorsWorks,
-    /// CollectMultipleErrors, etc.) all chain through these getters after the collect call returns.
-    /// </summary>
-    private static void WireNavCurrentThreadSession(object skeletonSession)
-    {
-        var nclAsm = AppDomain.CurrentDomain.GetAssemblies()
-            .FirstOrDefault(a => a.GetName().Name == "Microsoft.Dynamics.Nav.Ncl");
-        if (nclAsm == null) return;
-        var tTLS = nclAsm.GetType("Microsoft.Dynamics.Nav.Runtime.NavThreadLocalStorage");
-        var tRef = nclAsm.GetType("Microsoft.Dynamics.Nav.Runtime.Reference`1");
-        var tNavSession = skeletonSession.GetType();
-        if (tTLS == null || tRef == null) return;
-
-        var pCurrent = tTLS.GetProperty("Current", BindingFlags.Public | BindingFlags.Static);
-        var current = pCurrent?.GetValue(null);
-        if (current == null) return;
-        var pSession = tTLS.GetProperty("Session", BindingFlags.Public | BindingFlags.Instance);
-        if (pSession == null) return;
-
-        // Already set?
-        if (pSession.GetValue(current) != null) return;
-
-        // Build Reference<NavSession>(_skeletonSession). The single-arg ctor is public.
-        var refClosed = tRef.MakeGenericType(tNavSession);
-        var refInstance = Activator.CreateInstance(refClosed, new[] { skeletonSession });
-        pSession.SetValue(current, refInstance);
-        Console.Error.WriteLine("[RecordPatches] NavCurrentThread.Session wired to skeleton");
-    }
-
-    private static void InjectSkeletonSystemCodeunitFactory(object skeletonSession)
-    {
-        var nclAsm = AppDomain.CurrentDomain.GetAssemblies()
-            .FirstOrDefault(a => a.GetName().Name == "Microsoft.Dynamics.Nav.Ncl");
-        if (nclAsm == null) return;
-
-        var tFactory = nclAsm.GetType("Microsoft.Dynamics.Nav.Runtime.NavSystemCodeunitFactory");
-        var tGlobalTriggers = nclAsm.GetType("Microsoft.Dynamics.Nav.Runtime.NavSystemCodeunitGlobalTriggers");
-        var tNavCompany = nclAsm.GetType("Microsoft.Dynamics.Nav.Runtime.NavCompany");
-        if (tFactory == null || tGlobalTriggers == null || tNavCompany == null) return;
-
-        // Get the skeleton company from the session.
-        var companyField = skeletonSession.GetType().GetField("company",
-            BindingFlags.NonPublic | BindingFlags.Instance);
-        var skeletonCompany = companyField?.GetValue(skeletonSession);
-        if (skeletonCompany == null) return;
-
-        // Build the factory with the REAL public ctor when the skeleton session is a
-        // TreeObject (it is — NavSession : TreeObject). A non-null `parent` makes the
-        // lazy trigger getters (ReportingTriggers etc.) construct real
-        // NavSystemCodeunit instances whose NavCodeunitHandle resolves through the
-        // runner's CreateTarget hooks — required for the report-execution chain
-        // (GetReportToRun → InvokeSubstituteReport, factory fork →
-        // InvokeApplicationReportMergeStrategy, custom merger → OnCustomDocumentMergerEx).
-        // Fall back to the old uninitialized skeleton if the ctor shape changed.
-        object factory;
-        var tTreeObject = nclAsm.GetType("Microsoft.Dynamics.Nav.Runtime.TreeObject");
-        var factoryCtor = tTreeObject != null
-            ? tFactory.GetConstructor(new[] { tTreeObject })
-            : null;
-        if (factoryCtor != null && tTreeObject!.IsInstanceOfType(skeletonSession))
-            factory = factoryCtor.Invoke(new[] { skeletonSession });
-        else
-            factory = RuntimeHelpers.GetUninitializedObject(tFactory);
-
-        // Build GlobalTriggers with its REAL public ctor, NavSystemCodeunitGlobalTriggers(
-        // TreeObject parent), parented to the session — same reasoning as the factory above.
-        //
-        // This used to be a GetUninitializedObject skeleton with `session` field-poked in.
-        // That skipped the base NavSystemCodeunit ctor, which is what allocates
-        // `codeunitHandle` — the handle BC's own NavGlobalTriggers.Insert/Modify/Delete/
-        // RenameAsync invoke the "Global Triggers" codeunit (2000000002) through. Without it
-        // no OnDatabase* / OnGlobal* event could ever be published, so AL subscribers to
-        // Codeunit::"Global Triggers" silently never fired (corpus CU60210). It also skipped
-        // the `triggersOnTables` field initializer, which the old code had to patch up by
-        // hand; the real ctor does both correctly.
-        object globalTriggers;
-        var gtCtor = tGlobalTriggers.GetConstructors(BindingFlags.Public | BindingFlags.Instance)
-            .FirstOrDefault(c => c.GetParameters().Length == 1
-                && c.GetParameters()[0].ParameterType.IsInstanceOfType(skeletonSession));
-        if (gtCtor != null)
-        {
-            globalTriggers = gtCtor.Invoke(new[] { skeletonSession });
-        }
-        else
-        {
-            // Ncl shape changed. Fall back to the old skeleton rather than crash, but say so
-            // loudly: on this path every global/database trigger stays silently undelivered.
-            Console.Error.WriteLine(
-                "[RecordPatches] WARN: NavSystemCodeunitGlobalTriggers(TreeObject) ctor not found — "
-                + "falling back to an uninitialized skeleton; AL subscribers to "
-                + "Codeunit::\"Global Triggers\" will not fire.");
-            globalTriggers = RuntimeHelpers.GetUninitializedObject(tGlobalTriggers);
-            var fSession = tGlobalTriggers.GetField("session",
-                BindingFlags.NonPublic | BindingFlags.Instance);
-            fSession?.SetValue(globalTriggers, skeletonSession);
-        }
-
-        // triggersOnTables: Dictionary<int, Triggers> — initialize empty so Monitor.TryEnter
-        // doesn't NRE on the locked object. The dict is normally allocated via field initializer
-        // which GetUninitializedObject skips.
-        var fTriggersOnTables = tGlobalTriggers.GetField("triggersOnTables",
-            BindingFlags.NonPublic | BindingFlags.Instance);
-        if (fTriggersOnTables != null && fTriggersOnTables.GetValue(globalTriggers) == null)
-        {
-            var dictType = fTriggersOnTables.FieldType;          // Dictionary<int, Triggers>
-            fTriggersOnTables.SetValue(globalTriggers, Activator.CreateInstance(dictType));
-        }
-
-        // GetTriggersOnTable is BC's own body again — it invokes GetDatabaseTableTriggerSetup
-        // on the Global Triggers codeunit, whose AL subscribers decide the per-table mask.
-
-        // Wire global triggers into factory.
-        var fGlobalTriggers = tFactory.GetField("globalTriggers",
-            BindingFlags.NonPublic | BindingFlags.Instance);
-        fGlobalTriggers?.SetValue(factory, globalTriggers);
-
-        // openedDialogRegistry — normally allocated by NavCompany's field initializer,
-        // skipped by GetUninitializedObject. NavOpenDialogTracking (report execution:
-        // RunReportInternalCoreAsync) pushes onto it → NRE without this.
-        var tDialogRegistry = nclAsm.GetType("Microsoft.Dynamics.Nav.Runtime.NavOpenedDialogRegistry");
-        var fDialogRegistry = tNavCompany.GetField("openedDialogRegistry",
-            BindingFlags.NonPublic | BindingFlags.Instance);
-        if (tDialogRegistry != null && fDialogRegistry != null
-            && fDialogRegistry.GetValue(skeletonCompany) == null)
-        {
-            fDialogRegistry.SetValue(skeletonCompany,
-                Activator.CreateInstance(tDialogRegistry, nonPublic: true));
-        }
-
-        // The skeleton company has no Tree: it was produced by GetUninitializedObject, so it
-        // never ran TreeObject's ctor and was never parented to anything. NavCompany.
-        // RegisterFormInternal builds `new NavFormHandle(this, form)` with the COMPANY as
-        // parent, and TreeHandler's ctor throws InvalidOperationException("Parent.Tree cannot
-        // be null") for a parent with no tree — so registering a form (which BC does before
-        // dispatching a modal page to its handler) failed outright.
-        //
-        // Parenting it to the session is where a real BC server puts it, and going through
-        // BC's own CreateTreeHandler means the handler computes its own `session` from the
-        // parent exactly as it would there.
-        var fTreeObjTree = nclAsm.GetType("Microsoft.Dynamics.Nav.Runtime.TreeObject")?
-            .GetField("tree", BindingFlags.NonPublic | BindingFlags.Instance);
-        if (fTreeObjTree != null && fTreeObjTree.GetValue(skeletonCompany) == null)
-        {
-            var createHandler = nclAsm.GetType("Microsoft.Dynamics.Nav.Runtime.TreeHandler")?
-                .GetMethod("CreateTreeHandler", BindingFlags.Public | BindingFlags.Static);
-            if (createHandler != null)
-            {
-                var companyTree = createHandler.Invoke(null, new[] { skeletonSession, skeletonCompany });
-                AlRunner.Infrastructure.FieldPoke.SetInstance(fTreeObjTree, skeletonCompany, companyTree!);
-            }
-            else
-            {
-                Console.Error.WriteLine(
-                    "[RecordPatches] TreeHandler.CreateTreeHandler NOT FOUND — the skeleton company keeps "
-                    + "a null Tree and RegisterForm will throw");
-            }
-        }
-
-        // registeredForms — same class of gap as openedDialogRegistry above: allocated by
-        // NavCompany's ctor (`registeredForms = new Dictionary<Guid, NavFormHandle>()`),
-        // which GetUninitializedObject skips. BC's modal-page dispatch registers the form
-        // before showing it and unregisters it in a finally, so `lock (registeredForms)`
-        // threw ArgumentNullException from inside Monitor.Enter — surfacing to AL as a bare
-        // "Value cannot be null." on the RunModal line, naming nothing.
-        //
-        // A real empty dictionary is the faithful value: a company that has no open forms
-        // is exactly the runner's state, and BC populates and drains it itself from here on.
-        var fRegisteredForms = tNavCompany.GetField("registeredForms",
-            BindingFlags.NonPublic | BindingFlags.Instance);
-        if (fRegisteredForms != null && fRegisteredForms.GetValue(skeletonCompany) == null)
-            fRegisteredForms.SetValue(skeletonCompany,
-                Activator.CreateInstance(fRegisteredForms.FieldType));
-
-        // Inject factory into NavCompany.SystemCodeunitFactory auto-property backing field.
-        var fFactory = tNavCompany.GetField("<SystemCodeunitFactory>k__BackingField",
-            BindingFlags.NonPublic | BindingFlags.Instance);
-        fFactory?.SetValue(skeletonCompany, factory);
-
-        // W-8a PR1: populate NavCompany.trackChanges so NavRecord.InsertAsync's
-        // `ParentCompany.TrackChanges.TrackChange(...)` call doesn't NRE on the getter.
-        // We leave the internal `trackedChanges` dictionary null — TrackChange checks
-        // `if (trackedChanges == null) return;` and exits cleanly.
-        // (Ref: Ncl NavTrackChanges.TrackChange.)
-        var tTrackChanges = nclAsm.GetType("Microsoft.Dynamics.Nav.Runtime.NavTrackChanges");
-        var fTrackChanges = tNavCompany.GetField("trackChanges",
-            BindingFlags.NonPublic | BindingFlags.Instance);
-        if (tTrackChanges != null && fTrackChanges != null && fTrackChanges.GetValue(skeletonCompany) == null)
-        {
-            // Prefer BC's REAL ctor, `NavTrackChanges(NavCompany parent)`. It parents the
-            // object into the tree and allocates `trackedChanges`, both of which are now
-            // required: NavCompany.RegisterFormInternal calls RegisterFormTableChanges, which
-            // does `lock (trackedChanges)` with no null guard — unlike TrackChange, which is
-            // why the GetUninitializedObject shell below was sufficient until form
-            // registration started happening. It only became constructible once the company
-            // itself got a Tree (above); before that its base ctor would have thrown.
-            object? trackChanges = null;
-            var realCtor = tTrackChanges.GetConstructor(
-                BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance,
-                binder: null, types: new[] { tNavCompany }, modifiers: null);
-            if (realCtor != null)
-            {
-                try { trackChanges = realCtor.Invoke(new[] { skeletonCompany }); }
-                catch (Exception ex)
-                {
-                    var inner = ex is TargetInvocationException tie ? tie.InnerException ?? ex : ex;
-                    Console.Error.WriteLine(
-                        $"[RecordPatches] NavTrackChanges ctor failed ({inner.GetType().Name}: {inner.Message}) "
-                        + "— falling back to an uninitialized shell; form registration will throw");
-                }
-            }
-            // Fallback keeps the previous behaviour rather than leaving the field null.
-            trackChanges ??= RuntimeHelpers.GetUninitializedObject(tTrackChanges);
-            fTrackChanges.SetValue(skeletonCompany, trackChanges);
-        }
     }
 
     // ─── Hook Implementations ───────────────────────────────────────────────────
@@ -1732,68 +1072,6 @@ public static partial class RecordPatches
         return das;
     }
 
-    private static void EnsureSkeletonDatabase(object session)
-    {
-        // _skeletonDatabase is pre-built in Register(); NavSession.Database is JMP-hooked to return it.
-        // Nothing to inject on the session object itself.
-    }
-
-    private static object? BuildSqlSortingProperties()
-    {
-        if (_tSqlSortingProperties == null) return null;
-        try
-        {
-            // SqlSortingProperties(CultureInfo culture, CompareOptions compareOptions, string collation)
-            var sortingPropsCtor = _tSqlSortingProperties.GetConstructors(
-                    BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.Instance)
-                .FirstOrDefault(c => {
-                    var ps = c.GetParameters();
-                    return ps.Length == 3
-                        && ps[0].ParameterType == typeof(System.Globalization.CultureInfo)
-                        && ps[1].ParameterType == typeof(System.Globalization.CompareOptions);
-                });
-            if (sortingPropsCtor == null) return null;
-            return sortingPropsCtor.Invoke(new object[] {
-                System.Globalization.CultureInfo.InvariantCulture,
-                System.Globalization.CompareOptions.IgnoreCase,
-                "Latin1_General_CI_AS"
-            });
-        }
-        catch (Exception ex)
-        {
-            Console.Error.WriteLine($"[RecordPatches] BuildSqlSortingProperties failed: {ex.Message}");
-            return null;
-        }
-    }
-
-    private static object? BuildCollationAwareComparer()
-    {
-        if (_tCollationAwareStringComparer == null || _tSqlSortingProperties == null) return null;
-        var sortingProps = _sqlSortingProperties ?? BuildSqlSortingProperties();
-        if (sortingProps == null) return null;
-        try
-        {
-            var compCtor = _tCollationAwareStringComparer
-                .GetConstructors(BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.Instance)
-                .FirstOrDefault(c => c.GetParameters().Length == 1
-                    && c.GetParameters()[0].ParameterType == _tSqlSortingProperties);
-            return compCtor?.Invoke(new[] { sortingProps });
-        }
-        catch (Exception ex)
-        {
-            Console.Error.WriteLine($"[RecordPatches] BuildCollationAwareComparer failed: {ex.Message}");
-            return null;
-        }
-    }
-
-    /// <summary>
-    /// Replacement for NavSession.get_SortingProperties — Database.SqlSortingProperties NREs because
-    /// the skeleton NavDatabase does not have a collation set up for the lazy-init path. Return the
-    /// pre-built SqlSortingProperties from RecordPatches.Register.
-    /// </summary>
-    [MethodImpl(MethodImplOptions.NoInlining)]
-    public static object? NavSession_get_SortingProperties(object self) => _sqlSortingProperties;
-
     /// <summary>
     /// Replacement for DataAccessSource.GetDataAccessForTable(NCLMetaTable, bool).
     /// Uses one shared in-memory provider per (DataAccessSource,table) for regular tables,
@@ -1875,654 +1153,6 @@ public static partial class RecordPatches
     }
 
     /// <summary>
-    /// The --test-data on-demand load, installed by TestDataProvisioner.Arm() and null for
-    /// every run that did not pass the flag (so a default run pays one null check per
-    /// first-touch of a table). A delegate rather than a direct call because this file is the
-    /// MECHANISM half: which tables a backup offers, and from which backup, is policy, and it
-    /// lives in TestDataProvisioner. Arguments are (DataAccessSource, tableId).
-    /// </summary>
-    internal static Action<object, int>? TestDataOnDemandLoader;
-
-    /// <summary>
-    /// Told the id of a table whose storage was published from inside ANOTHER table's hydration
-    /// and therefore could not be loaded there (#2877). Installed by TestDataProvisioner.Arm()
-    /// alongside the loader, and null for every run without --test-data.
-    ///
-    /// The point is reporting, not mechanism: without it the provisioner has no record for that
-    /// table at all, so TableOutcome answers null — which under the on-demand policy means
-    /// "nothing in this run ever touched it", the opposite of what happened. Same argument as
-    /// #2240.
-    /// </summary>
-    internal static Action<int>? TestDataDeferredLoadNotifier;
-
-    /// <summary>
-    /// Told (table id, reason) when a deferred load could not be run after all, because the
-    /// store had rows by then or could not be read. Arriving here means the table ends the run
-    /// WITHOUT its backup rows, so it is reported rather than skipped — a silent skip is what
-    /// produced #2877. See .claude/rules/loud-failures.md.
-    /// </summary>
-    internal static Action<int, string>? TestDataDeferredLoadWriteOffNotifier;
-
-    /// <summary>Re-entrancy depth for the loader — NOT a "which tables are loaded" cache,
-    /// which #2262 rules out on purpose.
-    ///
-    /// Hydrating a table runs BC's own metadata and NavValue construction, and that code can
-    /// reach a Record of ANOTHER table, which lands back in GetDataAccessForTableCore and
-    /// would recurse.
-    ///
-    /// Skipping the nested load is only safe because the omission is RECORDED. The nested call
-    /// has already published the nested table's storage by the time the load is refused, and
-    /// "storage presence IS the have-we-loaded-this answer" then made the omission permanent:
-    /// every later touch found the entry and never loaded it, so the table silently kept none
-    /// of its backup rows for the whole run (#2877). GetOrCreateHydratedDataAccessCore's nested
-    /// branch marks that instance as owing a load, and the next touch outside a materialisation
-    /// settles it — see RecordPatches.TableMaterialisation.cs.</summary>
-    [ThreadStatic] private static int _testDataLoadDepth;
-
-    private static void InvokeTestDataOnDemandLoader(object source, int tableId)
-    {
-        var loader = TestDataOnDemandLoader;
-        if (loader == null || _testDataLoadDepth > 0) return;
-        _testDataLoadDepth++;
-        try { loader(source, tableId); }
-        finally { _testDataLoadDepth--; }
-    }
-
-    public static object NavDataAccessSource_GetDataAccessForTable(object self, NCLMetaTable table, bool isTemporary)
-    {
-        var dataAccess = GetDataAccessForTableCore(self, table, isTemporary);
-
-        // Both branches below land on a TempTableDataProvider, so the provider alone
-        // cannot say whether it is standing in for SQL or genuinely serving a
-        // `temporary` record — and the two shapes disagree about whether an
-        // uncommitted BLOB write reaches the stored row (corpus 60940, issue #1751).
-        // This is the one place that still knows, so record it here.
-        // A TableType = CRM table served by BC's CrmTestDataProvider is, in BC too, a
-        // temp-token DataAccess over a TempTableDataProvider (CrmTableConnection.CreateDataAccess
-        // passes DataAccessTableVersionTokens.CreateForTempTable) — not SQL-backed.
-        if (!isTemporary && !TableConnectionPatches.IsExternalTableType(table, out _))
-            BlobStoreIsolationPatches.MarkDatabaseBacked(dataAccess);
-
-        return dataAccess;
-    }
-
-    private static object GetDataAccessForTableCore(object self, NCLMetaTable table, bool isTemporary)
-    {
-        try
-        {
-            if (isTemporary)
-            {
-                // A `temporary` record gets a fresh, private store and NONE of the
-                // virtual-table populates below. Register it so the populates that run later
-                // (at find/Get time, keyed only on table id) honour the same invariant this
-                // early return does -- issue #2524.
-                var tempDataAccess = _mCreateTempDataAccess!.Invoke(self, new object[] { table })!;
-                _temporaryRecordDataAccess.AddOrUpdate(tempDataAccess, _temporaryRecordSentinel);
-                return tempDataAccess;
-            }
-
-            // Per-(DataAccessSource, tableId) cache so Insert+Find on the same regular table
-            // share storage.
-            var perTable = _dataAccessByTable.GetValue(self,
-                static _ => new ConcurrentDictionary<int, object>());
-            var tableId = table.TableId;
-
-            // ── External table types (CRM / ExternalSQL / Exchange / MicrosoftGraph) ─────
-            // BC's own GetDataAccessForTable switches on table.TableType here and asks the
-            // session's TableConnectionManager for the CURRENT connection of that type; the
-            // connection builds the DataAccess. Every one of these used to fall through to the
-            // temp store below as if it were a Normal table — a silent fake — because the
-            // metadata layer mapped every TableType to Normal. With '@@test@@' registered the
-            // CRM branch is BC's CrmTestDataProvider (Guid PK auto-assigned on insert); an
-            // unregistered type is BC's own "not registered" error; a live connection is
-            // refused by name. Not cached in perTable: the connection owns one provider per
-            // table id itself (CrmTableConnection.testDataProviders), exactly as on a service
-            // tier, and Unregister must drop it. See TableConnectionPatches (#2725).
-            if (TableConnectionPatches.IsExternalTableType(table, out var externalTableType))
-            {
-                var externalSession = _fDasSession?.GetValue(self)
-                    ?? throw new AlRunner.Infrastructure.RunnerOutOfScopeException(
-                        $"Record {tableId} (TableType = {externalTableType})",
-                        "table-connections — DataAccessSource has no skeleton session; see docs/scope.md",
-                        "table-connections");
-                return TableConnectionPatches.GetExternalDataAccess(
-                    externalSession, table, externalTableType, _fDasGlobalFilters?.GetValue(self));
-            }
-
-            // ── Virtual Field system table (2000000041) ──────────────────────────────────
-            // The Field table is virtual: the service tier computes its rows on the fly from
-            // NCLMetadata (one row per NCLMetaField of the filtered TableNo) via the native
-            // FieldDataProvider. Routing it to our empty in-memory store returns zero rows,
-            // which makes BC code that iterates it (e.g. "Library - Workflow".EnableWorkflow,
-            // Field.SetRange(TableNo,<t>); Field.FindSet()) throw "There is no Field within the
-            // filter."
-            //
-            // The block below builds a managed Field-row provider: it populates our in-memory
-            // store with REAL Field rows produced by BC's OWN managed row-builder
-            // FieldDataProvider.GetFieldRecordBuffer (a pure NCLMetaField→NavValue[] projection,
-            // NOT the crashing native find path) — see RecordPatches.FieldVirtualTable.cs. The
-            // populate is faithful and works (hundreds of rows insert cleanly for every table).
-            //
-            // DEFAULT-ON. The subsequent `Field.FindSet()` is routed through a managed find
-            // interception (a guard prepended to DataAccess.InnerFindAsync — see
-            // RecordPatches.FieldFindIntercept.cs) that, for 2000000041 ONLY, runs the find
-            // entirely in managed code (provider.Find → ResultSet → ResultSetEnumerator),
-            // bypassing BC's native InnerFindAsync SQL transactional-cache prologue which AVs on
-            // this virtual system table (its per-object SystemId/PK caches + table-version tokens
-            // are never allocated — crash file-proven even with zero rows and TableType=Temporary).
-            // The filtered TableNo's field rows are populated on demand at find time. Every other
-            // table falls through to the original native InnerFindAsync unchanged.
-            if (IsFieldVirtualTable(table))
-            {
-                if (!perTable.TryGetValue(tableId, out var fieldDa))
-                {
-                    var created = _mCreateTempDataAccess!.Invoke(self, new object[] { table })!;
-                    fieldDa = perTable.GetOrAdd(tableId, created);
-                }
-                // Top up rows for every source table currently materialised (idempotent). The
-                // subsequent Field.FindSet() is routed through DataAccess_FindAsync (a managed
-                // bypass of BC's R2R InnerFindAsync, which AVs on this virtual system table) —
-                // see RecordPatches.FieldFindIntercept.cs — and the filtered TableNo is populated
-                // on demand there. This whole path is now DEFAULT-ON (no env gate): the find
-                // interception means a populated Field table no longer crashes under R2R.
-                var session = _fDasSession?.GetValue(self)
-                    ?? throw FieldVirtualShapeGap("DataAccessSource has no skeleton session");
-                PopulateFieldVirtualTable(fieldDa, table, session);
-                return fieldDa;
-            }
-
-            // ── AllObj system virtual table (2000000038) ─────────────────────────────────
-            // Also virtual on the service tier (AllObjDataProvider computes rows from
-            // NCLMetadata.GetSnapshotOfAllObjects, whose body we Cecil-neuter because it
-            // NREs on the skeleton NCLMetadata). Routed to the same in-memory store as
-            // every other table, but POPULATED with one row per object the runner knows
-            // about, so AllObj.Get(<type>, <id>) and filtered iteration answer truthfully
-            // instead of always returning nothing.
-            // See RecordPatches.AllObjVirtualTable.cs.
-            if (IsAllObjVirtualTable(table))
-            {
-                if (!perTable.TryGetValue(tableId, out var allObjDa))
-                {
-                    var createdAllObj = _mCreateTempDataAccess!.Invoke(self, new object[] { table })!;
-                    allObjDa = perTable.GetOrAdd(tableId, createdAllObj);
-                }
-                PopulateAllObjVirtualTable(allObjDa, table);
-                return allObjDa;
-            }
-
-            // ── AllObjWithCaption system virtual table (2000000058) ──────────────────────
-            // AllObj plus the Object Caption column, virtual for the same reason, and the
-            // documented way for AL to render an object's caption (lookup pages bound to
-            // it, TableRelation, and lookup(...) CalcFormula FlowFields). Same rows and
-            // same key as AllObj — the inventory is literally shared, so the two tables
-            // cannot disagree about which objects exist.
-            // See RecordPatches.AllObjWithCaptionVirtualTable.cs.
-            if (IsAllObjWithCaptionVirtualTable(table))
-            {
-                if (!perTable.TryGetValue(tableId, out var allObjCaptionDa))
-                {
-                    var createdAllObjCaption = _mCreateTempDataAccess!.Invoke(self, new object[] { table })!;
-                    allObjCaptionDa = perTable.GetOrAdd(tableId, createdAllObjCaption);
-                }
-                PopulateAllObjWithCaptionVirtualTable(allObjCaptionDa, table);
-                return allObjCaptionDa;
-            }
-
-            // ── Integer system virtual table (2000000026) ────────────────────────────────
-            // Served by BC's OWN IntegerDataProvider, through BC's own
-            // DataAccessSource.GetVirtualDataAccess — rows are computed per request and none
-            // are stored, so an open-ended range answers on BC's ±1e9 clamp instead of from a
-            // materialised window (#3485). Not cached in perTable: BC caches it per table id
-            // on the DataAccessSource itself. See RecordPatches.IntegerVirtualTable.cs.
-            if (IsIntegerVirtualTable(table))
-            {
-                return GetIntegerVirtualDataAccess(self, table);
-            }
-
-            // ── All Profile system virtual table (2000000178) ────────────────────────────
-            // Virtual on the service tier too: AllProfileDataProvider's rows are every
-            // profile every published app declares plus the tenant-owned ones. It is the
-            // table Profile List / Profile Card are bound to and the one
-            // Conf./Personalization Mgt. resolves a user's role centre through, so an empty
-            // store made every read of it raise "There is no All Profile within the filter."
-            // Populated once per provider — unlike AllObj this table is WRITTEN by AL, and a
-            // top-up on a later handout would resurrect a just-deleted row.
-            // See RecordPatches.AllProfileVirtualTable.cs.
-            if (IsAllProfileVirtualTable(table))
-            {
-                if (!perTable.TryGetValue(tableId, out var allProfileDa))
-                {
-                    var createdAllProfile = _mCreateTempDataAccess!.Invoke(self, new object[] { table })!;
-                    allProfileDa = perTable.GetOrAdd(tableId, createdAllProfile);
-                }
-                PopulateAllProfileVirtualTable(allProfileDa, table);
-                return allProfileDa;
-            }
-
-            // ── Date system virtual table (2000000007) ───────────────────────────────────
-            // Served by BC's OWN DateDataProvider, through BC's own
-            // DataAccessSource.GetVirtualDataAccess — one row per period is computed per
-            // request and none are stored, so a range open at one end runs out to BC's own
-            // first/last period start instead of stopping at a materialised window (#3506).
-            // Not cached in perTable: BC caches it per table id on the DataAccessSource
-            // itself. See RecordPatches.DateVirtualTable.cs.
-            if (IsDateVirtualTable(table))
-            {
-                return GetDateVirtualDataAccess(self, table);
-            }
-
-            // ── Report Layout List system virtual table (2000000234) ─────────────────────
-            // Virtual on the service tier too (its rows are the layouts every published
-            // app declares, plus tenant layouts). BC's own by-name layout resolution
-            // (ReportLayoutSelection.GetLayoutByNameAndAppIDAsync) reads exactly this
-            // table, so populating it from the compiler-captured `rendering { layout(…) }`
-            // declarations makes selection-by-name work through BC's own code path.
-            // See RecordPatches.ReportLayoutListVirtualTable.cs.
-            if (IsReportLayoutListVirtualTable(table))
-            {
-                if (!perTable.TryGetValue(tableId, out var layoutDa))
-                {
-                    var createdLayout = _mCreateTempDataAccess!.Invoke(self, new object[] { table })!;
-                    layoutDa = perTable.GetOrAdd(tableId, createdLayout);
-                }
-                PopulateReportLayoutListVirtualTable(layoutDa, table);
-                return layoutDa;
-            }
-
-            // ── Report Metadata (2000000139) / Report Data Items (2000000203) ────────────
-            // Virtual on the service tier too: their rows are computed from the metadata of
-            // every published report. They are the documented way for AL to discover a
-            // report's caption, request-page flag and dataset shape without running it, and
-            // an empty store makes every such lookup answer "no such report" / "no data
-            // items". See RecordPatches.ReportMetadataVirtualTable.cs.
-            if (IsReportMetadataVirtualTable(table))
-            {
-                if (!perTable.TryGetValue(tableId, out var reportMetaDa))
-                {
-                    var createdReportMeta = _mCreateTempDataAccess!.Invoke(self, new object[] { table })!;
-                    reportMetaDa = perTable.GetOrAdd(tableId, createdReportMeta);
-                }
-                PopulateReportMetadataVirtualTable(reportMetaDa, table);
-                return reportMetaDa;
-            }
-
-            // ── Metadata Permission Set system virtual table (2000000250) ───────────────
-            // Virtual on the service tier too (MetadataPermissionSetDataProvider computes
-            // its rows from the permission sets the installed apps declare). An empty store
-            // makes Microsoft's own "Users - Create Super User" (codeunit 9000) fail its
-            // `MetadataPermissionSet.Get(<null guid>, 'SUPER')`, so every AL test that
-            // creates a user dies before it starts (issue #2313).
-            // See RecordPatches.MetadataPermissionSetVirtualTable.cs.
-            if (IsMetadataPermissionSetVirtualTable(table))
-            {
-                if (!perTable.TryGetValue(tableId, out var permSetDa))
-                {
-                    var createdPermSet = _mCreateTempDataAccess!.Invoke(self, new object[] { table })!;
-                    permSetDa = perTable.GetOrAdd(tableId, createdPermSet);
-                }
-                PopulateMetadataPermissionSetVirtualTable(permSetDa, table);
-                return permSetDa;
-            }
-
-            // ── Permission Set system table (2000000004) ────────────────────────────────
-            // Declared as an ordinary (obsolete-pending) table, but a modern tier does not
-            // read it from SQL: DataAccessSource.GetVirtualDataProvider routes it to
-            // PermissionSetDataProvider, which computes one row per ASSIGNABLE permission
-            // set the installed apps declare. An empty store made Microsoft's own
-            // "Permissions Mock"(codeunit 131006).Assign fail its PermissionSet.Get(RoleID),
-            // which is every named entry point of "Library - Lower Permissions" (issue
-            // #3344). See RecordPatches.PermissionSetSystemTable.cs.
-            if (IsPermissionSetSystemTable(table))
-            {
-                if (!perTable.TryGetValue(tableId, out var permSetSysDa))
-                {
-                    var createdPermSetSys = _mCreateTempDataAccess!.Invoke(self, new object[] { table })!;
-                    permSetSysDa = perTable.GetOrAdd(tableId, createdPermSetSys);
-                }
-                // The assignable-only shape — what walking the table lists. A keyed Get()
-                // repopulates with the non-assignable sets as well, through
-                // DataAccess_PermissionSetSystemTableGuardForGet.
-                PopulatePermissionSetSystemTable(permSetSysDa, table, includeNonAssignable: false);
-                return permSetSysDa;
-            }
-
-            // ── Permission system virtual table (2000000005) ────────────────────────────
-            // Virtual on the service tier too: PermissionDataProvider computes one row per
-            // permission a permission set grants, from the same metadata inventory the two
-            // tables above read. An empty store made every `Permission.Get(role, type, id)`
-            // answer "does not exist" and every filtered walk come back empty, while the
-            // set's HEADER listed correctly one table over — so a set's grants were
-            // unreadable with nothing reporting a gap (#3695).
-            // See RecordPatches.PermissionSystemTable.cs.
-            if (IsPermissionSystemTable(table))
-            {
-                if (!perTable.TryGetValue(tableId, out var permDa))
-                {
-                    var createdPerm = _mCreateTempDataAccess!.Invoke(self, new object[] { table })!;
-                    permDa = perTable.GetOrAdd(tableId, createdPerm);
-                }
-                var permSession = _fDasSession?.GetValue(self)
-                    ?? throw PermissionSystemTableShapeGap(
-                        "DataAccessSource has no skeleton session, so BC's own "
-                        + "PermissionDataProvider cannot be constructed");
-                PopulatePermissionSystemTable(permDa, table, permSession);
-                return permDa;
-            }
-
-            // ── Aggregate Permission Set system virtual table (2000000167) ──────────────
-            // Virtual on the service tier too: its rows are the UNION of System-scope
-            // (Metadata Permission Set, 2000000250 — just above) and Tenant-scope (Tenant
-            // Permission Set, 2000000165) rows, computed by BC's own
-            // AggregatePermissionSetDataProvider, driven directly by reflection. An empty
-            // store made every `Record "Aggregate Permission Set".Get(...)` fail "does not
-            // exist" — the root of a 14-test "already bound" cascade in Microsoft's own
-            // Tests-SINGLESERVER bucket (issue #2357, ruled out as a binding-mechanism
-            // defect by #2393). See RecordPatches.AggregatePermissionSetVirtualTable.cs.
-            if (IsAggregatePermissionSetVirtualTable(table))
-            {
-                if (!perTable.TryGetValue(tableId, out var aggPermSetDa))
-                {
-                    var createdAggPermSet = _mCreateTempDataAccess!.Invoke(self, new object[] { table })!;
-                    aggPermSetDa = perTable.GetOrAdd(tableId, createdAggPermSet);
-                }
-                var aggPermSetSession = _fDasSession?.GetValue(self)
-                    ?? throw AggregatePermissionSetShapeGap(
-                        "DataAccessSource has no skeleton session, so BC's own "
-                        + "AggregatePermissionSetDataProvider cannot be constructed");
-                PopulateAggregatePermissionSetVirtualTable(aggPermSetDa, table, aggPermSetSession);
-                return aggPermSetDa;
-            }
-
-            if (IsReportDataItemsVirtualTable(table))
-            {
-                if (!perTable.TryGetValue(tableId, out var reportDiDa))
-                {
-                    var createdReportDi = _mCreateTempDataAccess!.Invoke(self, new object[] { table })!;
-                    reportDiDa = perTable.GetOrAdd(tableId, createdReportDi);
-                }
-                PopulateReportDataItemsVirtualTable(reportDiDa, table);
-                return reportDiDa;
-            }
-
-            // ── Table Metadata (2000000136) ──────────────────────────────────────────────
-            // Virtual on the service tier too: one row per table in the application. An
-            // empty store makes every lookup answer "no such table", which is what broke
-            // Base App "Page Management".GetDefaultLookupPageID on custom tables.
-            // See RecordPatches.TableMetadataVirtualTable.cs.
-            if (IsTableMetadataVirtualTable(table))
-            {
-                if (!perTable.TryGetValue(tableId, out var tableMetaDa))
-                {
-                    var createdTableMeta = _mCreateTempDataAccess!.Invoke(self, new object[] { table })!;
-                    tableMetaDa = perTable.GetOrAdd(tableId, createdTableMeta);
-                }
-                PopulateTableMetadataVirtualTable(tableMetaDa, table);
-                return tableMetaDa;
-            }
-
-            // ── Page Metadata (2000000138) ───────────────────────────────────────────────
-            // Virtual on the service tier too: one row per page in the application. An
-            // empty store makes every lookup answer "no such page", which is what broke
-            // Base App "Page Management".GetDefaultCardPageID's SourceTable+PageType scan
-            // fallback for tables declaring no LookupPageId. See
-            // RecordPatches.PageMetadataVirtualTable.cs (#1769).
-            if (IsPageMetadataVirtualTable(table))
-            {
-                if (!perTable.TryGetValue(tableId, out var pageMetaDa))
-                {
-                    var createdPageMeta = _mCreateTempDataAccess!.Invoke(self, new object[] { table })!;
-                    pageMetaDa = perTable.GetOrAdd(tableId, createdPageMeta);
-                }
-                PopulatePageMetadataVirtualTable(pageMetaDa, table);
-                return pageMetaDa;
-            }
-
-            // ── Time Zone (2000000164) ───────────────────────────────────────────────────
-            // Virtual on the service tier too (TimeZoneDataProvider enumerates the HOST's
-            // TimeZoneInfo.GetSystemTimeZones() and numbers them 1..N). An empty store made
-            // every read answer "no such time zone" silently. The runner enumerates the same
-            // host call BC does, which on Linux means IANA ids where a Windows-hosted tier
-            // reports Windows ids — a deliberate, permanent divergence recorded in
-            // docs/limitations.md. See RecordPatches.TimeZoneVirtualTable.cs (#2584).
-            if (IsTimeZoneVirtualTable(table))
-            {
-                if (!perTable.TryGetValue(tableId, out var timeZoneDa))
-                {
-                    var createdTimeZone = _mCreateTempDataAccess!.Invoke(self, new object[] { table })!;
-                    timeZoneDa = perTable.GetOrAdd(tableId, createdTimeZone);
-                }
-                PopulateTimeZoneVirtualTable(timeZoneDa, table);
-                return timeZoneDa;
-            }
-
-            // ── Session (2000000009) ─────────────────────────────────────────────────────
-            // Virtual on the service tier too, and SessionDataProvider returns exactly ONE
-            // row — the reading session, My Session = true — not one per logged-on user.
-            // Routed to the same in-memory store as every other virtual table here and
-            // populated with that one row, read back from the skeleton NavSession so the
-            // table cannot disagree with SessionId() / UserId(). An empty store made every
-            // read answer "nobody is logged on", which is a wrong answer rather than a
-            // missing one. See RecordPatches.SessionVirtualTable.cs (#2940).
-            if (IsSessionVirtualTable(table))
-            {
-                if (!perTable.TryGetValue(tableId, out var sessionDa))
-                {
-                    var createdSession = _mCreateTempDataAccess!.Invoke(self, new object[] { table })!;
-                    sessionDa = perTable.GetOrAdd(tableId, createdSession);
-                }
-                var sessionTableSession = _fDasSession?.GetValue(self)
-                    ?? throw SessionVirtualShapeGap(
-                        "DataAccessSource has no skeleton session, so there is no session "
-                        + "identity to read the row back from");
-                PopulateSessionVirtualTable(sessionDa, table, sessionTableSession);
-                return sessionDa;
-            }
-
-            // ── Feature Key (2000000211) ─────────────────────────────────────────────────
-            // Routed to BC's OWN FeatureKeyDataProvider: its feature list is a hardcoded static
-            // in Microsoft.Dynamics.Nav.Types, so the rows are BC's rather than a second copy
-            // that would drift. An empty store made Base Application's Feature Management read
-            // every feature as unregistered and silently win the legacy code path.
-            // NOT a claim about any specific feature's shipped state: the 28.1 set measured
-            // here is 14 features, every one State = None. An earlier version of this comment
-            // said CalcOnlyVisibleFlowFields ships AllUsers (ON); that was read off the wrong
-            // Types.dll and is false for 28.1/28.4. Measure BuildFeatureKeys() against the
-            // artifact under test before asserting any feature's state.
-            // POPULATED READ-ONLY: real BC's Modify writes new state through to table
-            // 2000000210, which is not implemented here, so a Modify would land in the temp
-            // store and go nowhere. Issue #2585 tracks the write path.
-            // See RecordPatches.FeatureKeyVirtualTable.cs (#2585).
-            if (IsFeatureKeyVirtualTable(table))
-            {
-                if (!perTable.TryGetValue(tableId, out var featureKeyDa))
-                {
-                    var createdFeatureKey = _mCreateTempDataAccess!.Invoke(self, new object[] { table })!;
-                    featureKeyDa = perTable.GetOrAdd(tableId, createdFeatureKey);
-                }
-                var featureKeySession = _fDasSession?.GetValue(self)
-                    ?? throw FeatureKeyShapeGap(
-                        "DataAccessSource has no skeleton session, so BC's own "
-                        + "FeatureKeyDataProvider cannot be constructed");
-                PopulateFeatureKeyVirtualTable(featureKeyDa, table, featureKeySession);
-                return featureKeyDa;
-            }
-
-            // ── Windows Language (2000000045) ────────────────────────────────────────────
-            // Virtual on the service tier too (WindowsLanguageDataProvider iterates BC's own
-            // WindowsLanguageHelper.AllCultures). An empty store made every language lookup
-            // answer "no such language" silently. The six license-derived columns and the four
-            // installed-resource columns throw instead of guessing — see
-            // RecordPatches.WindowsLanguageVirtualTable.cs (#2581).
-            if (IsWindowsLanguageVirtualTable(table))
-            {
-                if (!perTable.TryGetValue(tableId, out var windowsLanguageDa))
-                {
-                    var createdWindowsLanguage = _mCreateTempDataAccess!.Invoke(self, new object[] { table })!;
-                    windowsLanguageDa = perTable.GetOrAdd(tableId, createdWindowsLanguage);
-                }
-                PopulateWindowsLanguageVirtualTable(windowsLanguageDa, table);
-                return windowsLanguageDa;
-            }
-
-            // ── NAV App Extra (2000000157) ───────────────────────────────────────────────
-            // Virtual on the service tier too (NavAppExtraDataProvider computes one row per
-            // app from the session's app metadata). An empty store made Published
-            // Application's "Tenant Visible" and "PerTenant Or Installed" Lookup FlowFields
-            // read FALSE for every app — the runner answering "no app is visible to this
-            // tenant", which is a wrong answer and not a missing one, and the same
-            // Boolean-default shape a real tier already contradicted once for the Installed
-            // FlowField next door (#3066). BC's own provider is tried first; the fallback
-            // builds the rows from the same loaded-module list and the same
-            // AppPackageIdentity values as the Published Application seeder, so the two
-            // tables cannot disagree about which app a runtime package id names.
-            // See RecordPatches.NavAppExtraVirtualTable.cs (#3072).
-            if (IsNavAppExtraVirtualTable(table))
-            {
-                if (!perTable.TryGetValue(tableId, out var navAppExtraDa))
-                {
-                    var createdNavAppExtra = _mCreateTempDataAccess!.Invoke(self, new object[] { table })!;
-                    navAppExtraDa = perTable.GetOrAdd(tableId, createdNavAppExtra);
-                }
-                var navAppExtraSession = _fDasSession?.GetValue(self)
-                    ?? throw NavAppExtraShapeGap(
-                        "DataAccessSource has no skeleton session, so BC's own "
-                        + "NavAppExtraDataProvider cannot be constructed");
-                PopulateNavAppExtraVirtualTable(navAppExtraDa, table, navAppExtraSession);
-                return navAppExtraDa;
-            }
-
-            // ── CodeUnit Metadata (2000000137) ───────────────────────────────────────────
-            // Virtual on the service tier too (CodeUnitDataProvider computes one row per
-            // codeunit from NCLMetadata). An empty store makes every lookup answer "no such
-            // codeunit", so Get() silently returns false and FindSet() raises — and a
-            // TableRelation to this table refuses a codeunit that really is in the run.
-            // The last missing member of the Table/Page/Report Metadata family above.
-            // See RecordPatches.CodeunitMetadataVirtualTable.cs (#2544).
-            if (IsCodeunitMetadataVirtualTable(table))
-            {
-                if (!perTable.TryGetValue(tableId, out var codeunitMetaDa))
-                {
-                    var createdCodeunitMeta = _mCreateTempDataAccess!.Invoke(self, new object[] { table })!;
-                    codeunitMetaDa = perTable.GetOrAdd(tableId, createdCodeunitMeta);
-                }
-                PopulateCodeunitMetadataVirtualTable(codeunitMetaDa, table);
-                return codeunitMetaDa;
-            }
-
-            // ── Object Metadata (2000000071) ─────────────────────────────────────────────
-            // NOT a virtual table: a real application-database system table, read with plain
-            // SQL by Ncl's own ObjectMetadataStorage. The runner has no application database,
-            // so its store was empty and a FindLast raised "There is no Object Metadata
-            // within the filter" — which is how it takes out Microsoft's own
-            // Codeunit136608.VerifyValidatePackageCodeunitFailed (#2519).
-            //
-            // Because the table IS real, a --test-data backup can genuinely carry rows for it.
-            // So the on-demand loader runs FIRST on a freshly created store, and the populator
-            // below does nothing when the store already holds a row: real rows win, synthesis
-            // is the fallback. Every other branch in this method serves a table no backup can
-            // ever have rows for, which is why only this one loads before populating.
-            //
-            // That precedence needs the hand-out to be ordered as well as the load, or a second
-            // thread is given the store between "created" and "hydrated", finds it empty and
-            // synthesises over rows that are about to arrive (#2788). GetOrCreateHydratedDataAccess
-            // is what guarantees it — see RecordPatches.TableMaterialisation.cs — so the populate
-            // below always runs on a store that is either hydrated or never will be.
-            // See RecordPatches.ObjectMetadataSystemTable.cs.
-            //
-            // A store published by a NESTED materialisation is the one case where the populate
-            // must NOT run yet: it owes a --test-data load that could not run there, and
-            // synthesising into it now would make that load a mix of real and synthesised rows.
-            // MaterialiseObjectMetadataStore holds the populate off until the debt is settled
-            // (#2877) — with no loader installed nothing is ever owed and this is unchanged.
-            if (IsObjectMetadataSystemTable(table))
-                return MaterialiseObjectMetadataStore(self, perTable, table, tableId);
-
-            // ── Object (2000000001) HAS NO BRANCH HERE, ON PURPOSE (#3071) ───────────────
-            // It used to have one, projecting the runner's object inventory into the legacy
-            // registry so that Object Metadata."Object ID"'s declared
-            // TableRelation = Object.ID WHERE(Type = FIELD("Object Type")) pointed at
-            // something (#2774). A real service tier then measured the table: corpus codeunit
-            // 61202 (StefanMaron/BusinessCentral.AL.Language.Tests#197) found it present,
-            // readable and EMPTY on seven BC OnPrem legs, with a control arm reading the
-            // populated sibling in the same session so "empty" could not be an unreadable
-            // table. A declared relation is not evidence that its target is populated.
-            //
-            // So this table now takes the generic fall-through at the end of this method, the
-            // one every other application-database table takes. That is deliberate rather than
-            // incidental: the fall-through is the SAME GetOrCreateHydratedDataAccess call the
-            // deleted branch made, so a --test-data backup's real rows still land, in the same
-            // order, with the same #2788 hand-out guarantee — only the synthesis is gone.
-            // See RecordPatches.ObjectSystemTable.cs for the measurement.
-
-            // ── Page Control Field (2000000192) ──────────────────────────────────────────
-            // Virtual on the service tier too: one row per field control declared on a
-            // page, INCLUDING controls declared Visible = false. An empty store made every
-            // filtered query answer "no rows" silently (no error), so a test asserting a
-            // control is absent would have passed against a broken provider too. See
-            // RecordPatches.PageControlFieldVirtualTable.cs (#1779).
-            if (IsPageControlFieldVirtualTable(table))
-            {
-                if (!perTable.TryGetValue(tableId, out var pageControlFieldDa))
-                {
-                    var createdPageControlField = _mCreateTempDataAccess!.Invoke(self, new object[] { table })!;
-                    pageControlFieldDa = perTable.GetOrAdd(tableId, createdPageControlField);
-                }
-                PopulatePageControlFieldVirtualTable(pageControlFieldDa, table);
-                return pageControlFieldDa;
-            }
-
-            // ── --test-data on-demand load (#2262) ───────────────────────────────────────
-            // A store that does not have this table yet is exactly when a --test-data run needs
-            // its rows read out of the backup. Same choke point the virtual tables above use,
-            // and for the same reason: it is the only place a table's storage is materialised,
-            // so the load always lands before the operation that triggered it — a read and a
-            // write are equally covered.
-            //
-            // Storage presence IS the "have we loaded this" answer, so there is no flag to
-            // keep in step: RestoreInstallBaselineSnapshot repopulates perTable from exactly
-            // the snapshot it restores, so a table the last restore carried is present and
-            // does not reach the create below. See TestDataProvisioner's header.
-            //
-            // Only the GetOrAdd winner loads — a loser would be hydrating into storage it is
-            // about to throw away — and no racer is handed the storage until that load has
-            // finished, so a caller here can never act on a half-hydrated table (#2788).
-            // See RecordPatches.TableMaterialisation.cs.
-            return GetOrCreateHydratedDataAccess(self, perTable, table, tableId);
-        }
-        catch (Exception ex)
-        {
-            var inner = ex is System.Reflection.TargetInvocationException tie ? tie.InnerException ?? ex : ex;
-            // Header + trace in ONE tagged write: Log.FilteredWriter matches per write and
-            // a bare stack trace has no `[Component]` tag, so a second call would print its
-            // frames at default verbosity under a header the filter had just dropped.
-            Console.Error.WriteLine(
-                $"[RecordPatches] GetDataAccessForTable failed for table "
-                + $"{table?.TableName ?? "null"}: {inner.GetType().Name}: {inner.Message}"
-                + $"\n{inner.StackTrace}");
-            throw;
-        }
-    }
-
-    /// <summary>
-    /// Replacement for NavDatabase.CollationAwareStringComparer getter.
-    /// Returns a CollationAwareStringComparer using InvariantCulture + IgnoreCase.
-    /// </summary>
-    [MethodImpl(MethodImplOptions.NoInlining)]
-    public static object? NavDatabase_get_CollationAwareStringComparer(NavDatabase self)
-    {
-        // Check if already populated on the skeleton instance (set by EnsureSkeletonDatabase).
-        if (_fNavDatabaseCollation != null)
-        {
-            var existing = _fNavDatabaseCollation.GetValue(self);
-            if (existing != null) return existing;
-        }
-        var built = BuildCollationAwareComparer();
-        if (built != null && _fNavDatabaseCollation != null)
-            _fNavDatabaseCollation.SetValue(self, built);
-        return built;
-    }
-
-    /// <summary>
     /// Replacement for NCLMetaApplicationObject.get_ApplicationObjectClrType.
     /// The real getter does lock(nclMetaObjectCLRTypeContainer) which NREs when the container
     /// is null (our CreateFromMetaTable-built tables never go through NCLCodeLoader).
@@ -2580,270 +1210,5 @@ public static partial class RecordPatches
     [MethodImpl(MethodImplOptions.NoInlining)]
     public static Guid NewSequentialId_Replacement()
         => Guid.NewGuid();
-
-    // ------------------------------------------------------------------
-    // NavTextConstant.get_Value — sync underbelly hit by every NavText(constant) ctor
-    // ------------------------------------------------------------------
-    //
-    // The real getter chains through `NavCurrentThread.ResolveAppGroup().GroupId` and
-    // `NavCurrentThread.Session.LocalLanguage / GlobalLanguage` to pick a language.
-    // NavCurrentThread.Session is null on the skeleton thread → NRE on every read of a
-    // NavTextConstant (which the AL emitter generates for every Label, including the
-    // five `TestFieldValidationCodeTxt`/`TestFieldCodeTxt`/etc. used by Assert codeunit).
-    //
-    // Empirically this is what causes `Assert.ExpectedTestFieldError` to NRE in Release
-    // mode: the AL-emit OnRun has expressions like
-    //     `NavTextExtensions.ALContains(this.lastErrorCode, new NavText(testFieldValidationCodeTxt))`
-    // and the `new NavText(constant)` invokes the implicit `NavStringValue → string`
-    // conversion which calls `NavTextConstant.Value` which NREs. Debug-mode emit happens
-    // to evaluate these in a different order that hides the NRE; Release-mode does not.
-    //
-    // Replace with a skeleton-safe lookup: pick the first English (LCID 1033) entry, or
-    // the first non-default entry, or empty. AL ships single-language ENU labels in v2,
-    // so the result is byte-identical to what the real getter returns under normal
-    // session state (LocalLanguage = 1033, fallback = 1033).
-
-    private static FieldInfo? _fNavTextConstant_multiLanguage;
-    private static FieldInfo? _fMultiLanguage_languageIds;
-    private static FieldInfo? _fMultiLanguage_texts;
-
-    /// <summary>
-    /// Replacement for NavStringValue.op_Implicit(NavStringValue → string). Original is
-    /// `value?.Value` — but `Value` on NavTextConstant NREs through NavCurrentThread.Session.
-    /// Route through our skeleton-safe NavTextConstant_get_Value when the input is a
-    /// NavTextConstant; otherwise read Value normally (other subtypes don't NRE).
-    /// </summary>
-    private static readonly System.Collections.Concurrent.ConcurrentDictionary<Type, PropertyInfo?> _pValueByType = new();
-
-    [MethodImpl(MethodImplOptions.NoInlining)]
-    public static string NavStringValue_op_Implicit(object? value)
-    {
-        if (value == null) return null!;
-        var t = value.GetType();
-        if (t.Name == "NavTextConstant")
-            return NavTextConstant_get_Value(value);
-        var prop = _pValueByType.GetOrAdd(t,
-            x => x.GetProperty("Value", BindingFlags.Public | BindingFlags.Instance));
-        return prop?.GetValue(value) as string ?? string.Empty;
-    }
-
-    [MethodImpl(MethodImplOptions.NoInlining)]
-    public static string NavTextConstant_get_Value(object self)
-    {
-        if (self == null) return string.Empty;
-        try
-        {
-            if (_fNavTextConstant_multiLanguage == null)
-                _fNavTextConstant_multiLanguage = self.GetType().GetField("multiLanguage",
-                    BindingFlags.NonPublic | BindingFlags.Instance);
-            var ml = _fNavTextConstant_multiLanguage?.GetValue(self);
-            if (ml == null) return string.Empty;
-            if (_fMultiLanguage_languageIds == null)
-                _fMultiLanguage_languageIds = ml.GetType().GetField("languageIds",
-                    BindingFlags.NonPublic | BindingFlags.Instance)
-                    ?? ml.GetType().GetField("LanguageIds",
-                        BindingFlags.NonPublic | BindingFlags.Instance);
-            if (_fMultiLanguage_texts == null)
-                _fMultiLanguage_texts = ml.GetType().GetField("texts",
-                    BindingFlags.NonPublic | BindingFlags.Instance)
-                    ?? ml.GetType().GetField("Texts",
-                        BindingFlags.NonPublic | BindingFlags.Instance);
-            // Try LanguageIds/Texts as properties (the public API on MultiLanguage).
-            var langProp = ml.GetType().GetProperty("LanguageIds",
-                BindingFlags.Public | BindingFlags.Instance);
-            var textProp = ml.GetType().GetProperty("Texts",
-                BindingFlags.Public | BindingFlags.Instance);
-            var langs = langProp?.GetValue(ml) ?? _fMultiLanguage_languageIds?.GetValue(ml);
-            var texts = textProp?.GetValue(ml) ?? _fMultiLanguage_texts?.GetValue(ml);
-            if (langs is System.Collections.IList ll && texts is System.Collections.IList tl && ll.Count > 0 && tl.Count > 0)
-            {
-                // Prefer English (1033) first, then any non-default.
-                for (int i = 0; i < ll.Count && i < tl.Count; i++)
-                {
-                    if (ll[i] is int lcid && lcid == 1033 && tl[i] is string s && !string.IsNullOrEmpty(s))
-                        return s;
-                }
-                if (tl[0] is string first) return first;
-            }
-        }
-        catch { }
-        return string.Empty;
-    }
-
-    // ------------------------------------------------------------------
-    // NavRecord.TestFieldNotBlank / TestFieldEquals / TestFieldError
-    // ------------------------------------------------------------------
-    //
-    // These three methods are the sync underbelly of `Rec.TestField(...)`.
-    // The real implementations format their failure message using:
-    //   - `base.Session.WindowsCulture`
-    //   - `ALFieldCaptionAsync(...).AsTask().GetAwaiter().GetResult()`
-    //   - `metaField.Parent.TableCaptionSafe`
-    //   - `PrimaryKeyString` (iterates key fields)
-    //   - `TryAddTestFieldAction(metaField)` (touches `Session.Diagnostics`,
-    //      `Session.Permissions`, `NavGlobal.NCLMetadata.GetMetaFormById` —
-    //      none of which the skeleton runtime initializes)
-    //
-    // Empirically (verified 2026-05-09 with Debug-mode emit) the throw path
-    // raises a NullReferenceException somewhere inside that argument list,
-    // which then surfaces with error code "NullReference" instead of
-    // "TestField". Assert.ExpectedTestFieldError sees the wrong code, calls
-    // its own Error path, and the test fails with NRE 12 times.
-    //
-    // Per HANDOFF §5.2 (Option C) we replace the throw path with a clean
-    // `NavTestFieldException.CreateNonblank/CreateMustBeEqualTo` call using
-    // safe arguments — same factory the real BC code uses, just with culture
-    // forced to InvariantCulture and PrimaryKeyValues="" to avoid the
-    // failing-skeleton property dives. The exception type is identical
-    // (`NavTestFieldException`) so `GetErrorCode` returns "TestField" and
-    // Assert.ExpectedTestFieldError's `LastErrorCode.Contains("TestField")`
-    // matches. The message contains the field caption, satisfying
-    // ExpectedTestFieldMessage's StrPos check.
-    //
-    // The pass path (value is non-blank / equal) is left to the real code:
-    // we delegate to it via reflection and only intercept the throw path.
-
-    private static MethodInfo? _mGetFieldValue;
-    private static readonly System.Collections.Concurrent.ConcurrentDictionary<Type, PropertyInfo?> _pIsZeroOrEmptyByType = new();
-    private static MethodInfo? _mNavTestFieldException_CreateNonblank;
-    private static MethodInfo? _mNavTestFieldException_CreateMustBeEqualTo;
-    private static PropertyInfo? _pNCLMetaFieldFieldNo;
-    private static PropertyInfo? _pNCLMetaFieldFieldName;
-    private static PropertyInfo? _pNCLMetaFieldParent;
-    private static PropertyInfo? _pNCLMetaTableTableName;
-
-    private static string SafeFieldName(object? metaField)
-    {
-        if (metaField == null) return string.Empty;
-        if (_pNCLMetaFieldFieldName == null)
-            _pNCLMetaFieldFieldName = metaField.GetType().GetProperty("FieldName",
-                BindingFlags.Public | BindingFlags.Instance);
-        return (string?)_pNCLMetaFieldFieldName?.GetValue(metaField) ?? string.Empty;
-    }
-
-    private static string SafeTableName(object? metaField)
-    {
-        if (metaField == null) return string.Empty;
-        if (_pNCLMetaFieldParent == null)
-            _pNCLMetaFieldParent = metaField.GetType().GetProperty("Parent",
-                BindingFlags.Public | BindingFlags.Instance);
-        var parent = _pNCLMetaFieldParent?.GetValue(metaField);
-        if (parent == null) return string.Empty;
-        if (_pNCLMetaTableTableName == null)
-            _pNCLMetaTableTableName = parent.GetType().GetProperty("TableName",
-                BindingFlags.Public | BindingFlags.Instance);
-        return (string?)_pNCLMetaTableTableName?.GetValue(parent) ?? string.Empty;
-    }
-
-    private static object CreateNavTestFieldException_Nonblank(object metaField)
-    {
-        if (_mNavTestFieldException_CreateNonblank == null)
-        {
-            var navTypes = AppDomain.CurrentDomain.GetAssemblies()
-                .FirstOrDefault(a => a.GetName().Name == "Microsoft.Dynamics.Nav.Types");
-            var t = navTypes?.GetType("Microsoft.Dynamics.Nav.Types.Exceptions.NavTestFieldException")
-                ?? navTypes?.GetType("Microsoft.Dynamics.Nav.Types.NavTestFieldException")
-                ?? navTypes?.GetTypes().FirstOrDefault(x => x.Name == "NavTestFieldException");
-            _mNavTestFieldException_CreateNonblank = t?.GetMethod("CreateNonblank",
-                BindingFlags.Public | BindingFlags.Static);
-        }
-        var m = _mNavTestFieldException_CreateNonblank
-            ?? throw new InvalidOperationException("NavTestFieldException.CreateNonblank not found");
-        // Signature: (CultureInfo, string fieldName, string tableName, string primaryKeyValues, ErrorInfoData=null)
-        var args = new object?[] { System.Globalization.CultureInfo.InvariantCulture,
-            SafeFieldName(metaField), SafeTableName(metaField), string.Empty, null };
-        return (Exception)m.Invoke(null, args)!;
-    }
-
-    private static object CreateNavTestFieldException_MustBeEqualTo(object metaField, string shouldBe, string current)
-    {
-        if (_mNavTestFieldException_CreateMustBeEqualTo == null)
-        {
-            var navTypes = AppDomain.CurrentDomain.GetAssemblies()
-                .FirstOrDefault(a => a.GetName().Name == "Microsoft.Dynamics.Nav.Types");
-            var t = navTypes?.GetType("Microsoft.Dynamics.Nav.Types.Exceptions.NavTestFieldException")
-                ?? navTypes?.GetType("Microsoft.Dynamics.Nav.Types.NavTestFieldException")
-                ?? navTypes?.GetTypes().FirstOrDefault(x => x.Name == "NavTestFieldException");
-            _mNavTestFieldException_CreateMustBeEqualTo = t?.GetMethod("CreateMustBeEqualTo",
-                BindingFlags.Public | BindingFlags.Static);
-        }
-        var m = _mNavTestFieldException_CreateMustBeEqualTo
-            ?? throw new InvalidOperationException("NavTestFieldException.CreateMustBeEqualTo not found");
-        // Signature: (CultureInfo, string fieldName, string tableName, string shouldBeValue,
-        //            string currentValue, string primaryKeyValues, ErrorInfoData=null)
-        var args = new object?[] { System.Globalization.CultureInfo.InvariantCulture,
-            SafeFieldName(metaField), SafeTableName(metaField), shouldBe, current, string.Empty, null };
-        return (Exception)m.Invoke(null, args)!;
-    }
-
-    private static bool TryGetFieldValueIsZeroOrEmpty(object navRecord, object metaField, out bool result)
-    {
-        result = true;
-        try
-        {
-            // Look up by the NCLMetaField base parameter type (the runtime metaField may be
-            // a derived subclass that the public method-resolution can't match directly).
-            if (_mGetFieldValue == null)
-            {
-                var nclMetaFieldT = metaField.GetType();
-                while (nclMetaFieldT != null && nclMetaFieldT.Name != "NCLMetaField")
-                    nclMetaFieldT = nclMetaFieldT.BaseType;
-                _mGetFieldValue = navRecord.GetType().GetMethod("GetFieldValue",
-                    BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.Instance,
-                    null, new[] { nclMetaFieldT ?? metaField.GetType() }, null);
-            }
-            var v = _mGetFieldValue?.Invoke(navRecord, new object[] { metaField });
-            if (v == null) { result = true; return true; }
-            // NavValue.IsZeroOrEmpty is virtual; cache the PropertyInfo per concrete subtype
-            // because the NCLMetaField argument can be different NavValue subclasses across
-            // calls (NavInteger / NavText / NavCode / …).
-            var p = _pIsZeroOrEmptyByType.GetOrAdd(v.GetType(),
-                t => t.GetProperty("IsZeroOrEmpty", BindingFlags.Public | BindingFlags.Instance));
-            var b = p?.GetValue(v) as bool?;
-            result = b ?? true;
-            return true;
-        }
-        catch { return false; }
-    }
-
-    /// <summary>
-    /// Replacement for NavRecord.TestFieldNotBlank(NCLMetaField, NavALErrorInfo).
-    /// Real method computes the error message via Session.WindowsCulture, async ALFieldCaption,
-    /// PrimaryKeyString, and TryAddTestFieldAction — all of which dereference skeleton state
-    /// that's null. Compute the not-blank predicate via the real `GetFieldValue/IsZeroOrEmpty`
-    /// (those work on a populated record) and on the throw path raise a NavTestFieldException
-    /// directly with InvariantCulture and minimal args, so error code is "TestField" and the
-    /// message contains the field caption — what Assert.ExpectedTestFieldError expects.
-    /// </summary>
-    [MethodImpl(MethodImplOptions.NoInlining)]
-    public static void NavRecord_TestFieldNotBlank(object self, object metaField, object? errorInfo)
-    {
-        if (metaField == null) throw new ArgumentNullException(nameof(metaField));
-        if (TryGetFieldValueIsZeroOrEmpty(self, metaField, out var isBlank) && !isBlank)
-            return; // value is set — nothing to assert.
-        throw (Exception)CreateNavTestFieldException_Nonblank(metaField);
-    }
-
-    /// <summary>
-    /// Replacement for NavRecord.TestFieldError(NCLMetaField, string, NavALErrorInfo).
-    /// Same rationale as TestFieldNotBlank — computes the error message safely.
-    /// </summary>
-    [MethodImpl(MethodImplOptions.NoInlining)]
-    public static void NavRecord_TestFieldError(object self, object metaField, string shouldBeValue, object? errorInfo)
-    {
-        if (metaField == null) throw new ArgumentNullException(nameof(metaField));
-        string current = "<N/A>";
-        try
-        {
-            if (_mGetFieldValue == null)
-                _mGetFieldValue = self.GetType().GetMethod("GetFieldValue",
-                    BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.Instance,
-                    null, new[] { metaField.GetType() }, null);
-            var v = _mGetFieldValue?.Invoke(self, new object[] { metaField });
-            current = v?.ToString() ?? "<N/A>";
-        }
-        catch { /* leave default */ }
-        throw (Exception)CreateNavTestFieldException_MustBeEqualTo(metaField, shouldBeValue ?? string.Empty, current);
-    }
 
 }
